Refactored the base KMW object; performed work to support minification.
