+++ web/samples/uncompiled.html
@@ -23,15 +23,7 @@
     </style> 
  
     <!-- Insert uncompiled KeymanWeb source scripts -->              
-    <script src="../source/kmwstring.js" type="application/javascript"></script>
-    <script src="../source/kmwbase.js" type="application/javascript"></script>
-    <script src="../source/keymanweb.js" type="application/javascript"></script>
-    <script src="../source/kmwosk.js" type="application/javascript"></script>
-    <script src="../source/kmwnative.js" type="application/javascript"></script>
-    <script src="../source/kmwcallback.js" type="application/javascript"></script>
-    <script src="../source/kmwkeymaps.js" type="application/javascript"></script>
-    <script src="../source/kmwlayout.js" type="application/javascript"></script>
-    <script src="../source/kmwinit.js" type="application/javascript"></script>             
+    <script src="../build/keymanweb.js" type="application/javascript"></script>           
 
     <!-- 
       For desktop browsers, a script for the user interface must be inserted here.
@@ -40,14 +32,13 @@
       The toolbar UI is best for any page designed to support keyboards for 
       a large number of languages.
     -->
-    <script src="../source/kmwuitoggle.js"></script>                                                                                           
+    <script src="../build/kmwuitoggle.js"></script>                                                                                           
  
     <!-- Initialization: set paths to keyboards, resources and fonts as required -->
     <script>
       var kmw=window.keyman;
       kmw.init({
-        attachType:'auto',
-        resources:'resources'
+        attachType:'auto'
         });
     </script>
